some feedback implemented
