+++ src/client/components/TimeZone/TimeZone.component.js
@@ -1,33 +1,32 @@
-import React from 'react';
+import React, { useState } from 'react';
 import DropDown from '../Dropdown/DropDown';
-import PropTypes, { shape } from 'prop-types';
 import 'moment-timezone';
 
-export const TimeZoneDropDown = (props) => {
+const timeZones = [
+  {
+    id: 1,
+    name: 'Europe/Copenhagen',
+  },
+  {
+    id: 2,
+    name: 'Europe/Dublin',
+  },
+  {
+    id: 3,
+    name: 'Europe/Budapest',
+  },
+  {
+    id: 4,
+    name: 'Australia/Sydney',
+  },
+];
+
+export const TimeZoneDropDown = () => {
+  const [name, setName] = useState('');
+
   return (
     <div>
-      <DropDown
-        items={props.timeZones}
-        value={props.name}
-        setValue={props.setName}
-      />
+      <DropDown items={timeZones} value={name} setValue={setName} />
     </div>
   );
-};
-
-TimeZoneDropDown.propTypes = {
-  timeZones: PropTypes.arrayOf(
-    shape({
-      id: PropTypes.number,
-      name: PropTypes.string,
-    }),
-  ),
-  name: PropTypes.string,
-  setName: PropTypes.func,
-};
-
-TimeZoneDropDown.defaultProps = {
-  timeZones: [],
-  name: '',
-  setName: null,
 };
